requirements.txt
albumentations==1.4.0
arabic_reshaper==3.0.0
python-bidi==0.4.2
comet_ml==3.38.1
coremltools==7.1
Flask==3.0.2
GitPython==3.1.42
ipython==8.12.3
jellyfish==1.0.3
keras==3.0.5
matplotlib==3.8.3
mss==9.0.1
nncf==2.8.1
numpy==1.26.4
onnx==1.15.0
onnxruntime==1.17.1
onnxsim==0.4.35
openvino==2023.3.0
pafy==0.5.5
pandas==2.2.1
persian==0.5.0
Pillow==10.2.0
psutil==5.9.8
pycocotools==2.0.7
PySide6==6.6.2
qtpy==2.4.1
PyYAML==6.0.1
PyQt6==6.6.1
Requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
setuptools==69.0.3
tensorflow==2.15.0.post1
tensorflowjs==4.17.0
tensorrt==8.6.1.post1
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
tritonclient==2.43.0
ultralytics==8.1.21
x2paddle==1.4.1


